Corrected int types to int54 to work with SQLITE db.
Made startup in program.cs return the database interface which has been populated by the correct database dictated by tablesconfig.json

+++ FileBox.Shared/Models/BaseFileInfo.cs
@@ -7,13 +7,13 @@ namespace FileBox.Shared.Models
 	{
 		[NameCast("BaseFileInfoId")]
 		[PropertyType("INTEGER", true)]
-		public int Id { get; set; }
+		public Int64 Id { get; set; }
 
 		[PropertyType("INTEGER")]
-		public int FileCode { get; set; }
+		public Int64 FileCode { get; set; }
 
 		[PropertyType("INTEGER")]
-		public int Size { get; set; }
+		public Int64 Size { get; set; }
 		public DateTime LastModified { get; set; }
 	}
 }
