+++ FileBox.Shared/Models/InfoValue.cs
@@ -7,10 +7,10 @@ namespace FileBox.Shared.Models
 	{
 		[NameCast("InfoValueId")]
 		[PropertyType("INTEGER", true)]
-		public int Id { get; set; }
+		public Int64 Id { get; set; }
 
 		[PropertyType("INTEGER")]
-		public int FileCode { get; set; }
+		public Int64 FileCode { get; set; }
 
 		[PropertyType("VARCHAR(24)")]
 		public string ValueName { get; set; }
